Extracted Grid DAO components into a separate file
