+++ backend/services/grid-storage/test/test-database.ts
@@ -4,7 +4,8 @@ const expect = chai.expect;
 chai.use(chai_as_promised);
 chai.should();
 
-import {connectGridDatabase, GridCRUDRepository, GridDAO, MongoGridCRUDRepository} from '../mongo-utils';
+import {connectGridDatabase} from '../mongo-utils';
+import {GridCRUDRepository, GridDAO, MongoGridCRUDRepository} from '../GridDAO'
 const DB_HOSTNAME_TEST : string = process.env.DB_HOSTNAME_TEST || '127.0.0.1'
 const DB_PORT_TEST : number = (process.env.DB_PORT_TEST) ? parseInt(process.env.DB_PORT_TEST) : 27017
 const testGrid : GridDAO = new GridDAO("Test",["ABC","DEF","GHI"]);
